+++ crates/arroyo-operator/src/lib.rs
@@ -21,6 +21,7 @@ pub mod connector;
 pub mod context;
 pub mod inq_reader;
 pub mod operator;
+pub mod udfs;
 
 pub trait TimerT: Data + PartialEq + Eq + 'static {}
