+++ crates/arroyo-worker/Cargo.toml
@@ -85,6 +85,9 @@ datafusion-common = "36.0"
 datafusion-execution = "36.0"
 base64 = "0.21.5"
 itertools = "0.12.0"
+async-ffi = "0.5.0"
+dlopen2 = "0.7.0"
+dlopen2_derive = "0.4.0"
 
 
 [dev-dependencies]
