--- crates/arroyo-worker/src/connectors/mod.rs
@@ -1,13 +1,0 @@
-pub mod blackhole;
-pub mod filesystem;
-pub mod fluvio;
-pub mod impulse;
-pub mod kafka;
-pub mod kinesis;
-pub mod mqtt;
-pub mod nexmark;
-pub mod polling_http;
-pub mod redis;
-pub mod two_phase_committer;
-pub mod webhook;
-pub mod websocket;
