--- crates/arroyo-worker/src/operators/mod.rs
@@ -1,2 +1,0 @@
-//pub mod async_map;
-pub mod watermark_generator;
